Move EmbeddedPlayers to better location and minor fixes
